Rename 'Deposit compliance' to 'REF 2021' (#365)
Adds more texts and descriptions about the REF. Links the section
to our REF audit guidance.

Adds an alert that the section is region-specific. The alert appears
on the Overview and in the top of the REF section for **non-UK**
repositories only.

Adds alert-outline icon needed for the alert.

Adds passing DataProvider's country code into templates.